Publicize functions in the autobumper
This enables reusing the functions as API for autobumping the
prow-images for other prow deployment, eg, [1].

[1]. https://github.com/openshift/release
