+++ experiment/autobumper/BUILD.bazel
@@ -6,10 +6,9 @@ go_library(
     importpath = "k8s.io/test-infra/experiment/autobumper",
     visibility = ["//visibility:private"],
     deps = [
-        "//experiment/image-bumper/bumper:go_default_library",
+        "//experiment/autobumper/bumper:go_default_library",
         "//prow/config/secret:go_default_library",
         "//prow/github:go_default_library",
-        "//robots/pr-creator/updater:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
     ],
 )
@@ -29,7 +28,10 @@ filegroup(
 
 filegroup(
     name = "all-srcs",
-    srcs = [":package-srcs"],
+    srcs = [
+        ":package-srcs",
+        "//experiment/autobumper/bumper:all-srcs",
+    ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
 )
