+++ experiment/autobumper/bumper/BUILD.bazel
@@ -1,0 +1,28 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["bumper.go"],
+    importpath = "k8s.io/test-infra/experiment/autobumper/bumper",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//experiment/image-bumper/bumper:go_default_library",
+        "//prow/github:go_default_library",
+        "//robots/pr-creator/updater:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
+    ],
+)
+
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [":package-srcs"],
+    tags = ["automanaged"],
+    visibility = ["//visibility:public"],
+)
